% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{addOutcomeNames}
\alias{addOutcomeNames}
\title{Add names to a data frame with outcome IDs}
\usage{
addOutcomeNames(data, outcomeIdColumnName = "outcomeId")
}
\arguments{
\item{data}{The data frame to add the outcome names to}

\item{outcomeIdColumnName}{The name of the column in the data frame that holds the outcome IDs.}
}
\description{
Add names to a data frame with outcome IDs
}

