% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EvaluatePredictiveModels.R
\name{evaluatePredictiveModels}
\alias{evaluatePredictiveModels}
\title{Compute evaluation metrics for the predictive models}
\usage{
evaluatePredictiveModels(outputFolder)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/)}
}
\description{
Compute evaluation metrics for the predictive models
}
\details{
This function computes the AUC and plots the ROC and calibration plots per predictive model.
}

