% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute OHDSI Celecoxib predictive modelsstudy}
\usage{
execute(connectionDetails, cdmDatabaseSchema,
  workDatabaseSchema = cdmDatabaseSchema,
  studyCohortTable = "ohdsi_celecoxib_prediction", oracleTempSchema = NULL,
  gap = 1, cdmVersion = 5, outputFolder, createCohorts = TRUE,
  createPredictiveModels = TRUE, packageResultsForSharing = TRUE,
  updateProgress = NULL)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the \code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include
both the database and schema name, for example 'cdm_data.dbo'.}

\item{workDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have write priviliges in this schema. Note that
for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.}

\item{studyCohortTable}{The name of the table that will be created in the work database schema. This table will hold the exposure and outcome
cohorts used in this study.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write priviliges for storing temporary tables.}

\item{cdmVersion}{Version of the CDM. Can be "4" or "5"}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/)}
}
\value{
TODO
}
\details{
This function executes the OHDSI Celecoxib predictive models study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(connectionDetails,
        cdmDatabaseSchema = "cdm_data",
        workDatabaseSchema = "results",
        oracleTempSchema = NULL,
        outputFolder = "c:/temp/study_results",
        cdmVersion = "5")

}

}

